// src/features/admin/doctors/profile/tabs/types.ts
import { LucideIcon } from "lucide-react";
import { ReactElement } from "react";

export type ScheduleItem = {
    day: string;
    hours: string;
  };
  
 export type ScheduleProps = {
    availability: ScheduleItem[];
    primaryColor?: string;
  };

  export type Service = {
    id: number;
    name: string;
    description: string;
    duration: string;
    price: string;
  };

 export  type ServicesProps = {
    services: Service[];
    primaryColor?: string;
  };

  export type TabList="about"|"services"|"schedule"

  export type tabHeaderType={
    activeTab:TabList
    onChange:(tab:TabList)=>void
  }

  export type Tab = {
    id: "about" | "services" | "schedule";
    label: string;
    icon: ReactElement
  };


export type AvailableTimeSlot = {
  id: string
  date: string
  time: string
  doctorId: string
  appointmentId: string | null
  createdAt: string
  updatedAt: string
  deleteAt: string | null
}

export type CreateAvailableTimeSlot = Omit<
  AvailableTimeSlot,
  'id' | 'createdAt' | 'updatedAt' | 'deleteAt'
>

export type UpdateAvailableTimeSlot = Partial<
  Omit<AvailableTimeSlot, 'id'>
> & {
  id: string
}